+++ plugins/template/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @bs-plugins/template
 
+## 0.2.14
+
+### Patch Changes
+
+- Added signed-double-submit-cookie CSRF check to csrf middleware
+- Updated dependencies
+  - @bs-core/shell@1.16.13
+
 ## 0.2.13
 
 ### Patch Changes