+++ shell/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bs-core/shell
 
+## 1.16.13
+
+### Patch Changes
+
+- Added signed-double-submit-cookie CSRF check to csrf middleware
+
 ## 1.16.12
 
 ### Patch Changes